cli/requirements.txt
wcwidth==0.1.9
win-unicode-console==0.5
sentry-sdk==0.19.4
tableformatter==0.1.4
pycryptodome==3.9.7
timeago==1.0.13
cmd2==1.0.1
appdirs==1.4.3
requests==2.23.0
psutil==5.7.0
pyzmq==19.0.0
selenium==3.141.0
flask
tqdm
pywin32
